# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首页
  /toRegister:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 注册页
  /user/createUser:
    post:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: rePassword
        type: string
      - description: 电话号码
        in: query
        name: phone
        type: string
      - description: 邮箱
        in: query
        name: email
        type: string
      responses:
        "200":
          description: message"}
          schema:
            type: string
      summary: 添加用户
      tags:
      - 用户模块
  /user/deleteUser:
    get:
      parameters:
      - description: 用户id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: message"}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/findUserByNameAndPwd:
    post:
      parameters:
      - description: 名字
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      responses:
        "200":
          description: message"}
          schema:
            type: string
      summary: 通过用户名和密码判断用户
      tags:
      - 用户模块
  /user/getUserList:
    get:
      responses:
        "200":
          description: message"}
          schema:
            type: string
      summary: 所有用户
      tags:
      - 用户模块
  /user/updateUser:
    post:
      parameters:
      - description: 用户id
        in: formData
        name: id
        type: string
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 密码
        in: formData
        name: password
        type: string
      - description: 电话号码
        in: formData
        name: phone
        type: string
      - description: 邮箱
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: message"}
          schema:
            type: string
      summary: 修改用户
      tags:
      - 用户模块
swagger: "2.0"